feat: add price display for rehab centers and improve admin interface
- Add min_price field to RehabCenter admin with collapsible section
- Change min_price from DecimalField to IntegerField for whole numbers
- Remove price display from clinic cards across all templates
- Replace duplicated card code with includes in index.html
- Update specialist detail page to show prices only for rehab centers
- Create migration for min_price field type change

+++ facilities/models.py
@@ -186,9 +186,7 @@ class RehabCenter(AbstractMedicalFacility):
         verbose_name=_('Длительность программы (дней)'),
         default=0
     )
-    min_price = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
+    min_price = models.IntegerField(
         verbose_name=_('Минимальная цена'),
         default=0
     )
